# openapi.yaml
openapi: 3.1.0
info:
  title: Trustify
  description: Software Supply-Chain Security API
  license:
    name: Apache License, Version 2.0
    identifier: Apache-2.0
  version: 0.4.0-beta.1
paths:
...
  /api/v2/sbom:
    get:
      tags:
      - sbom
      summary: List SBOMs
      operationId: listSboms
      parameters:
      - name: q
        in: query
        description: |
          EBNF grammar for the _q_ parameter:
          ```text
              q = ( values | filter ) { '&' q }
              values = value { '|', values }
              filter = field, operator, values
              operator = "=" | "!=" | "~" | "!~" | ">=" | ">" | "<=" | "<"
              value = (* any text but escape special characters with '\' *)
              field = (* must match an entity attribute name *)
          ```
          Any values in a _q_ will result in a case-insensitive "full
          text search", effectively producing an OR clause of LIKE
          clauses for every string-ish field in the resource being
          queried.

          Examples:
          - `foo` - any field containing 'foo'
          - `foo|bar` - any field containing either 'foo' OR 'bar'
          - `foo&bar` - some field contains 'foo' AND some field contains 'bar'

          A _filter_ may also be used to constrain the results. The
          filter's field name must correspond to one of the resource's
          attributes. If it doesn't, an error will be returned
          containing a list of the valid fields for that resource.

          The value 'null' is treated specially for [Not]Equal filters:
          it returns resources on which the field isn't set. Use the
          LIKE operator, `~`, to match a literal "null" string. Omit the
          value to match an empty string.

          Examples:
          - `name=foo` - entity's _name_ matches 'foo' exactly
          - `name~foo` - entity's _name_ contains 'foo', case-insensitive
          - `name~foo|bar` - entity's _name_ contains either 'foo' OR 'bar', case-insensitive
          - `name=` - entity's _name_ is the empty string, ''
          - `name=null` - entity's _name_ isn't set
          - `published>3 days ago` - date values can be "human time"

          Multiple full text searches and/or filters should be
          '&'-delimited -- they are logically AND'd together.

          - `red hat|fedora&labels:type=cve|osv&published>last wednesday 17:00`

          Fields corresponding to JSON objects in the database may use a
          ':' to delimit the column name and the object key,
          e.g. `purl:qualifiers:type=pom`

          Any operator or special character, e.g. '|', '&', within a
          value should be escaped by prefixing it with a backslash.
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: |
          EBNF grammar for the _sort_ parameter:
          ```text
              sort = field [ ':', order ] { ',' sort }
              order = ( "asc" | "desc" )
              field = (* must match the name of entity's attributes *)
          ```
          The optional _order_ should be one of "asc" or "desc". If
          omitted, the order defaults to "asc".

          Each _field_ name must correspond to one of the columns of the
          table holding the entities being queried. Those corresponding
          to JSON objects in the database may use a ':' to delimit the
          column name and the object key,
          e.g. `purl:qualifiers:type:desc`
        required: false
        schema:
          type: string
      - name: offset
        in: query
        description: |-
          The first item to return, skipping all that come before it.

          NOTE: The order of items is defined by the API being called.
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: limit
        in: query
        description: |-
          The maximum number of entries to return.

          Zero means: no limit
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: group
        in: query
        description: Filter by group IDs
        required: false
        schema:
          type: array
          items:
            type: string
            format: uuid
      responses:
        '200':
          description: Matching SBOMs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResults_SbomSummary'
    post:
      tags:
      - sbom
      summary: Upload a new SBOM
      operationId: uploadSbom
      parameters:
      - name: labels
        in: query
        description: |-
          Optional labels.

          Only use keys with a prefix of `labels.`
        required: true
        schema:
          $ref: '#/components/schemas/Labels'
      - name: format
        in: query
        description: The format of the uploaded document.
        required: false
        schema:
          type: string
          enum:
          - osv
          - csaf
          - cve
          - spdx
          - cyclonedx
          - clearlydefinedcuration
          - clearlydefined
          - cwecatalog
          - advisory
          - sbom
          - unknown
      - name: cache
        in: query
        description: Await loading the document into the analysis graph cache
        required: false
        schema:
          type: string
          enum:
          - skip
          - queue
          - wait
      - name: groups
        in: query
        description: Optional group IDs to assign the SBOM to
        required: false
        schema:
          type: array
          items:
            type: string
            format: uuid
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: array
              items:
                type: integer
                format: int32
                minimum: 0
        required: true
      responses:
        '201':
          description: Upload an SBOM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResult'
        '400':
          description: The file could not be parsed as an SBOM
...
  /api/v2/sbom/by-package:
    get:
      tags:
      - sbom
      summary: Find all SBOMs containing the provided package.
      description: |-
        The package can be provided either via a PURL or using the ID of a package as returned by
        other APIs, but not both.
      operationId: listRelatedSboms
      parameters:
      - name: q
        in: query
        description: |
          EBNF grammar for the _q_ parameter:
          ```text
              q = ( values | filter ) { '&' q }
              values = value { '|', values }
              filter = field, operator, values
              operator = "=" | "!=" | "~" | "!~" | ">=" | ">" | "<=" | "<"
              value = (* any text but escape special characters with '\' *)
              field = (* must match an entity attribute name *)
          ```
          Any values in a _q_ will result in a case-insensitive "full
          text search", effectively producing an OR clause of LIKE
          clauses for every string-ish field in the resource being
          queried.

          Examples:
          - `foo` - any field containing 'foo'
          - `foo|bar` - any field containing either 'foo' OR 'bar'
          - `foo&bar` - some field contains 'foo' AND some field contains 'bar'

          A _filter_ may also be used to constrain the results. The
          filter's field name must correspond to one of the resource's
          attributes. If it doesn't, an error will be returned
          containing a list of the valid fields for that resource.

          The value 'null' is treated specially for [Not]Equal filters:
          it returns resources on which the field isn't set. Use the
          LIKE operator, `~`, to match a literal "null" string. Omit the
          value to match an empty string.

          Examples:
          - `name=foo` - entity's _name_ matches 'foo' exactly
          - `name~foo` - entity's _name_ contains 'foo', case-insensitive
          - `name~foo|bar` - entity's _name_ contains either 'foo' OR 'bar', case-insensitive
          - `name=` - entity's _name_ is the empty string, ''
          - `name=null` - entity's _name_ isn't set
          - `published>3 days ago` - date values can be "human time"

          Multiple full text searches and/or filters should be
          '&'-delimited -- they are logically AND'd together.

          - `red hat|fedora&labels:type=cve|osv&published>last wednesday 17:00`

          Fields corresponding to JSON objects in the database may use a
          ':' to delimit the column name and the object key,
          e.g. `purl:qualifiers:type=pom`

          Any operator or special character, e.g. '|', '&', within a
          value should be escaped by prefixing it with a backslash.
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: |
          EBNF grammar for the _sort_ parameter:
          ```text
              sort = field [ ':', order ] { ',' sort }
              order = ( "asc" | "desc" )
              field = (* must match the name of entity's attributes *)
          ```
          The optional _order_ should be one of "asc" or "desc". If
          omitted, the order defaults to "asc".

          Each _field_ name must correspond to one of the columns of the
          table holding the entities being queried. Those corresponding
          to JSON objects in the database may use a ':' to delimit the
          column name and the object key,
          e.g. `purl:qualifiers:type:desc`
        required: false
        schema:
          type: string
      - name: offset
        in: query
        description: |-
          The first item to return, skipping all that come before it.

          NOTE: The order of items is defined by the API being called.
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: limit
        in: query
        description: |-
          The maximum number of entries to return.

          Zero means: no limit
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: purl
        in: query
        description: Find by PURL
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Purl'
      - name: cpe
        in: query
        description: Find by CPE
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Cpe'
      responses:
        '200':
          description: Matching SBOMs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResults_SbomSummary'
  /api/v2/sbom/count-by-package:
    get:
      tags:
      - sbom
      summary: Count all SBOMs containing the provided packages.
      description: |-
        The packages can be provided either via a PURL or using the ID of a package as returned by
        other APIs, but not both.
      operationId: countRelatedSboms
      parameters:
      - name: purl
        in: path
        description: Find by PURL
        required: true
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Purl'
      - name: cpe
        in: path
        description: Find by CPE
        required: true
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Cpe'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ExternalReferenceQuery'
        required: true
      responses:
        '200':
          description: Number of matching SBOMs per package
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  format: int64
  /api/v2/sbom/{id}:
    get:
      tags:
      - sbom
      summary: Get information about an SBOM
      operationId: getSbom
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: Matching SBOM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SbomSummary'
        '404':
          description: The SBOM could not be found
    delete:
      tags:
      - sbom
      summary: Delete an SBOM
      operationId: deleteSbom
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: Matching SBOM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SbomSummary'
        '404':
          description: The SBOM could not be found
  /api/v2/sbom/{id}/advisory:
    get:
      tags:
      - sbom
      summary: Get advisories for an SBOM
      operationId: getSbomAdvisories
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: Matching SBOM
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SbomAdvisory'
        '404':
          description: The SBOM could not be found
  /api/v2/sbom/{id}/all-license-ids:
    get:
      tags:
      - sbom
      operationId: listAllLicenseIds
      parameters:
      - name: id
        in: path
        description: ID of the SBOM to get the license IDs for
        required: true
        schema:
          type: string
      responses:
        '200':
          description: fetch all unique license id and license info id
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LicenseRefMapping'
        '400':
          description: Invalid UUID format.
...
  /api/v2/sbom/{id}/license-export:
    get:
      tags:
      - sbom
      operationId: getLicenseExport
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: license gzip files
          content:
            application/gzip:
              schema:
                type: array
                items:
                  type: integer
                  format: int32
                  minimum: 0
        '404':
          description: The document could not be found
...
  /api/v2/sbom/{key}/download:
    get:
      tags:
      - sbom
      summary: Download an SBOM
      operationId: downloadSbom
      parameters:
      - name: key
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: Download a an SBOM
          content:
            application/json:
              schema:
                type: string
                format: binary
        '404':
          description: The document could not be found
...
components:
  schemas:
...
    Cpe:
      type: string
      format: uri
...
    ExternalReferenceQuery:
      type: object
      properties:
        cpe:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Cpe'
            description: Find by CPE
        purl:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Purl'
            description: Find by PURL
...
    Id:
      type: string
      description: |-
        Identifier to a document, prefixed with the ID type.

        Either an internal ID of the document with the `urn:uuid:` scheme. Or using a digest, with the digest prefix. For example, `sha256:`.
      examples:
      - urn:uuid:018123ef-a791-40d8-b62a-f70a350245d4
      - sha256:dc60aeb735c16a71b6fc56e84ddb8193e3a6d1ef0b7e958d77e78fc039a5d04e
...
    IngestResult:
      type: object
      description: The result of the ingestion process
      required:
      - id
      properties:
        document_id:
          type:
          - string
          - 'null'
          description: The ID declared by the document
        id:
          $ref: '#/components/schemas/Id'
          description: The internal ID of the document
        warnings:
          type: array
          items:
            type: string
          description: Warnings that occurred during the import process
    Labels:
      type: object
      additionalProperties:
        type: string
...
    LicenseRefMapping:
      type: object
      required:
      - license_id
      - license_name
      properties:
        license_id:
          type: string
        license_name:
          type: string
...
    PaginatedResults_SbomSummary:
      type: object
      required:
      - items
      - total
      properties:
        items:
          type: array
          items:
            allOf:
            - $ref: '#/components/schemas/SbomHead'
            - $ref: '#/components/schemas/SourceDocument'
            - type: object
              required:
              - described_by
              properties:
                described_by:
                  type: array
                  items:
                    $ref: '#/components/schemas/SbomPackage'
        total:
          type: integer
          format: int64
          minimum: 0
...
    Purl:
      type: string
      format: uri
...
    SbomAdvisory:
      allOf:
      - $ref: '#/components/schemas/AdvisoryHead'
      - type: object
        required:
        - status
        properties:
          status:
            type: array
            items:
              $ref: '#/components/schemas/SbomStatus'
...
    SbomSummary:
      allOf:
      - $ref: '#/components/schemas/SbomHead'
      - $ref: '#/components/schemas/SourceDocument'
      - type: object
        required:
        - described_by
        properties:
          described_by:
            type: array
            items:
              $ref: '#/components/schemas/SbomPackage'
...
    SourceDocument:
      type: object
      required:
      - sha256
      - sha384
      - sha512
      - size
      - ingested
      properties:
        ingested:
          type: string
          format: date-time
          description: The timestamp the document was ingested
        sha256:
          type: string
        sha384:
          type: string
        sha512:
          type: string
        size:
          type: integer
          format: int64
          minimum: 0
...